fixed ui bug refactored react-hook-form handling into custom hook
